Seperate migrate function to own file. Function that controlls which migrations should be ran. Duplicate temp controller function
