+++ src/main/kotlin/no/risc/risc/RiScService.kt
@@ -4,14 +4,8 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.coroutineScope
-import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.JsonArray
-import kotlinx.serialization.json.JsonObject
-import kotlinx.serialization.json.JsonPrimitive
-import kotlinx.serialization.json.jsonArray
 import kotlinx.serialization.json.jsonObject
 import kotlinx.serialization.json.jsonPrimitive
 import no.risc.encryption.CryptoServiceIntegration
@@ -29,6 +23,8 @@ import no.risc.infra.connector.models.AccessTokens
 import no.risc.infra.connector.models.GCPAccessToken
 import no.risc.risc.models.RiScWrapperObject
 import no.risc.risc.models.UserInfo
+import no.risc.utils.migrateFrom33To40
+import no.risc.utils.migrateTo32To33
 import no.risc.utils.removePathRegex
 import no.risc.validation.JSONValidator
 import org.apache.commons.lang3.RandomStringUtils
@@ -125,6 +121,7 @@ class RiScService(
         owner: String,
         repository: String,
         accessTokens: AccessTokens,
+        latestSupportedVersion: String,
     ): List<RiScContentResultDTO> =
         coroutineScope {
             val msId =
@@ -156,7 +153,7 @@ class RiScService(
                                                  *
                                                  * In case a repository does not delete branches after merging pull requests,
                                                  * our state flowchart would go from SentForApproval (pull request) to
-                                                 * Draft (branch exists). Therefore we check if the content in Draft is equal to
+                                                 * Draft (branch exists). Therefore, we check if the content in Draft is equal to
                                                  * the content on the default branch (often main). If they are equal then we
                                                  * know we have a ros branch without changes, and therefore it should be in a
                                                  * Published state.
@@ -172,9 +169,9 @@ class RiScService(
                                                     )
 
                                                 val fileFound = published.status === GithubStatus.Success
-                                                val fileisEqual = published.data.equals(it.data)
+                                                val fileIsEqual = published.data.equals(it.data)
                                                 // Check if file exists and its content is equal.
-                                                if (fileFound && fileisEqual) {
+                                                if (fileFound && fileIsEqual) {
                                                     // Set its status to Published.
                                                     id.status = RiScStatus.Published
                                                 }
@@ -191,8 +188,7 @@ class RiScService(
                                         accessTokens.gcpAccessToken,
                                         id.pullRequestUrl,
                                     )
-                                    .let { migrateToNewMinor(it) }
-                                    .let { migrateFrom33To40(it) }
+                                    .let { migrate(it, latestSupportedVersion) }
                             } catch (e: Exception) {
                                 RiScContentResultDTO(
                                     riScId = id.id,
@@ -210,120 +206,33 @@ class RiScService(
             msRiSc.value
         }
 
-    // Update RiSc scenarios from schemaVersion 3.2 to 3.3. This is necessary because 3.3 is backwards compatible,
-    // and modifications can only be made when the schemaVersion is 3.3.
-    private fun migrateToNewMinor(obj: RiScContentResultDTO): RiScContentResultDTO {
+    private fun migrate(
+        obj: RiScContentResultDTO,
+        latestSupportedVersion: String,
+    ): RiScContentResultDTO {
         if (obj.riScContent == null) {
             return obj
         }
 
-        val migratedSchemaVersion = obj.riScContent.replace("\"schemaVersion\": \"3.2\"", "\"schemaVersion\": \"3.3\"")
-        return obj.copy(riScContent = migratedSchemaVersion)
-    }
-
-    /**
-     * Update RiSc content from version 3.3 to 4.0. Includes breaking changes.
-     *
-     * Changes include:
-     * - Bump schemaVersion to 4.0
-     *
-     * Replace values in vulnerabilities:
-     * - User repudiation -> Unmonitored use
-     * - Compromised admin user -> Unauthorized access
-     * - Escalation of rights -> Unauthorized access
-     * - Disclosed secret -> Information leak
-     * - Denial of service -> Excessive use
-     *
-     * Remove "owner" and "deadline" from actions
-     * Remove "existingActions" from scenarios
-     */
-    @OptIn(ExperimentalSerializationApi::class)
-    private fun migrateFrom33To40(obj: RiScContentResultDTO): RiScContentResultDTO {
-        if (obj.riScContent == null) {
-            return obj
-        }
-
-        var content = obj.riScContent
+        val content = obj.riScContent
 
         val json = Json { ignoreUnknownKeys = true }
         val jsonObject = json.parseToJsonElement(content).jsonObject.toMutableMap()
 
-        // Check if schemaVersion is 3.3, early return the object as it is if not
-        if (jsonObject["schemaVersion"]?.jsonPrimitive?.content != "3.3") {
+        val schemaVersion = jsonObject["schemaVersion"]?.jsonPrimitive?.content
+
+        if (schemaVersion == null || schemaVersion == latestSupportedVersion) {
             return obj
         }
 
-        // Replace schemaVersion
-        jsonObject["schemaVersion"] = JsonPrimitive("4.0")
-
-        // Update scenarios
-        val scenarios = jsonObject["scenarios"]?.jsonArray ?: return obj.copy(riScContent = content)
-        val updatedScenarios =
-            scenarios.map { scenario ->
-                val scenarioObject = scenario.jsonObject.toMutableMap()
-
-                // Remove "existingActions"
-                val scenarioDetails = scenarioObject["scenario"]?.jsonObject?.toMutableMap()
-                scenarioDetails?.remove("existingActions")
-
-                // Replace values in vulnerabilities array
-                val vulnerabilitiesArray = scenarioDetails?.get("vulnerabilities")?.jsonArray?.toMutableList()
-                if (vulnerabilitiesArray != null) {
-                    val replacementMap =
-                        mapOf(
-                            "User repudiation" to "Unmonitored use",
-                            "Compromised admin user" to "Unauthorized access",
-                            "Escalation of rights" to "Unauthorized access",
-                            "Disclosed secret" to "Information leak",
-                            "Denial of service" to "Excessive use",
-                        )
-
-                    val newVulnerabilities =
-                        vulnerabilitiesArray
-                            .map { it.jsonPrimitive.content }
-                            .toMutableSet()
-
-                    replacementMap.forEach { (oldValue, newValue) ->
-                        if (newVulnerabilities.contains(oldValue)) {
-                            newVulnerabilities.remove(oldValue)
-                            newVulnerabilities.add(newValue)
-                        }
-                    }
-
-                    // Convert back to JsonArray
-                    val updatedVulnerabilitiesArray = JsonArray(newVulnerabilities.map { JsonPrimitive(it) })
-                    scenarioDetails["vulnerabilities"] = updatedVulnerabilitiesArray
-                }
-
-                // Remove "owner" and "deadline" from actions
-                val actionsArray = scenarioDetails?.get("actions")?.jsonArray?.toMutableList()
-                actionsArray?.forEachIndexed { index, actionElement ->
-                    val actionObject = actionElement.jsonObject.toMutableMap()
-                    actionObject["action"]?.jsonObject?.toMutableMap()?.apply {
-                        this.remove("owner")
-                        this.remove("deadline")
-                    }?.let { updatedAction ->
-                        actionObject["action"] = JsonObject(updatedAction)
-                    }
-                    actionsArray[index] = JsonObject(actionObject)
-                }
-                actionsArray?.let { scenarioDetails["actions"] = JsonArray(it) }
-
-                scenarioDetails?.let { scenarioObject["scenario"] = JsonObject(it) }
-                JsonObject(scenarioObject)
+        val nextVersionObj =
+            when (schemaVersion) {
+                "3.2" -> migrateTo32To33(obj)
+                "3.3" -> migrateFrom33To40(obj)
+                else -> return obj
             }
 
-        jsonObject["scenarios"] = JsonArray(updatedScenarios)
-
-        // Convert the updated JSON object back to string with pretty printing
-        val prettyJson =
-            Json {
-                prettyPrint = true
-                prettyPrintIndent = "    "
-            }
-        content = prettyJson.encodeToString(JsonObject(jsonObject))
-
-        return obj.copy(riScContent = content, migrationChanges = true)
+        return migrate(nextVersionObj, latestSupportedVersion)
     }
 
     private fun GithubContentResponse.responseToRiScResult(
